requirements.txt
Django==1.2.5
pygments==1.4
pyYAML==3.09
